+++ app/src/androidTest/java/com/github/HumanLearning2021/HumanLearningApp/view/dataset_editing/CategoriesEditingActivityTest.kt
@@ -1,6 +1,9 @@
 package com.github.HumanLearning2021.HumanLearningApp.view.dataset_editing
 
+import android.content.Context
 import android.content.Intent
+import android.net.Uri
+import androidx.test.core.app.ApplicationProvider
 import androidx.test.espresso.Espresso.onView
 import androidx.test.espresso.action.ViewActions.click
 import androidx.test.espresso.action.ViewActions.typeText
@@ -25,6 +28,8 @@ import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
+import java.io.File
+import java.util.*
 import javax.inject.Inject
 
 
@@ -59,6 +64,23 @@ class CategoriesEditingActivityTest {
                     found = true
                 }
             }
+            if (!found) {
+                val cat = dbManagement.putCategory("${UUID.randomUUID()}")
+                dataset = dbManagement.putDataset("${UUID.randomUUID()}", setOf(cat))
+                val tmp = File.createTempFile("droid", ".png")
+                try {
+                    ApplicationProvider.getApplicationContext<Context>().resources.openRawResource(R.drawable.fork).use { img ->
+                        tmp.outputStream().use {
+                            img.copyTo(it)
+                        }
+                    }
+                    val uri = Uri.fromFile(tmp)
+                    dbManagement.putPicture(uri, cat)
+                } finally {
+                    tmp.delete()
+                }
+            }
+
             categories = dataset.categories
             nbCategories = categories.size
             datasetId = dataset.id as String
