+++ app/src/main/java/com/github/HumanLearning2021/HumanLearningApp/view/GoogleSignInWidget.kt
@@ -33,11 +33,14 @@ class GoogleSignInWidget : Fragment() {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        view.findViewById<Button>(R.id.loginButton).setOnClickListener(this::onLoginButtonPress)
+        view.findViewById<Button>(R.id.loginButton).setOnClickListener {
+            onLoginButtonPress()
+        }
         updateUi()
     }
 
-    fun onLoginButtonPress(view: View) {
+    private fun onLoginButtonPress() {
+        @Suppress("DEPRECATION")  // FIXME: use something non-deprecated
         startActivityForResult(
             presenter.intentForStartActivityForResult(),
             RC_SIGN_IN
@@ -46,6 +49,7 @@ class GoogleSignInWidget : Fragment() {
 
 
     override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
+        @Suppress("DEPRECATION")  // FIXME: use something non-deprecated
         super.onActivityResult(requestCode, resultCode, data)
 
         when (requestCode) {
