+++ app/src/main/java/com/github/HumanLearning2021/HumanLearningApp/view/MainActivity.kt
@@ -14,6 +14,15 @@ class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
+
+
+        if (savedInstanceState == null) {
+            val fragment = MainFragment()
+            supportFragmentManager
+                .beginTransaction()
+                .add(R.id.main_content, fragment)
+                .commit()
+        }
     }
 
     fun launchToDisplayDatasetActivity(view: View) {
