+++ app/src/main/java/com/github/HumanLearning2021/HumanLearningApp/view/MainFragment.kt
@@ -1,0 +1,19 @@
+package com.github.HumanLearning2021.HumanLearningApp.view
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.github.HumanLearning2021.HumanLearningApp.R
+
+class MainFragment : Fragment() {
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        return inflater.inflate(R.layout.fragment_main, container, false)
+    }
+
+}
