+++ app/src/main/res/layout/fragment_learning.xml
@@ -1,0 +1,71 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:context=".view.learning.LearningFragment">
+
+    <ImageView
+        android:id="@+id/learning_im_to_sort"
+        android:layout_width="@dimen/learning_image_size"
+        android:layout_height="@dimen/learning_image_size"
+        android:layout_margin="32dp"
+        android:background="@color/white"
+        android:contentDescription="@string/learning_im_to_sort_descr"
+        android:longClickable="true"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.5"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/constraintLayout"
+        app:srcCompat="@color/blue" />
+
+    <androidx.constraintlayout.widget.ConstraintLayout
+        android:id="@+id/constraintLayout"
+        android:layout_width="match_parent"
+        android:layout_height="0dp"
+        app:layout_constraintBottom_toTopOf="@+id/learning_im_to_sort"
+        app:layout_constraintTop_toTopOf="parent">
+
+        <ImageView
+            android:id="@+id/learning_cat_0"
+            android:layout_width="@dimen/learning_image_size"
+            android:layout_height="@dimen/learning_image_size"
+            android:background="@color/white"
+            android:contentDescription="@string/learning_cat_0_descr"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintEnd_toStartOf="@+id/learning_cat_1"
+            app:layout_constraintHorizontal_bias="0.5"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toTopOf="parent"
+            app:srcCompat="@color/red" />
+
+        <ImageView
+            android:id="@+id/learning_cat_1"
+            android:layout_width="@dimen/learning_image_size"
+            android:layout_height="@dimen/learning_image_size"
+            android:background="@color/white"
+            android:contentDescription="@string/learning_cat_1_descr"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintEnd_toStartOf="@+id/learning_cat_2"
+            app:layout_constraintHorizontal_bias="0.5"
+            app:layout_constraintStart_toEndOf="@+id/learning_cat_0"
+            app:layout_constraintTop_toTopOf="parent"
+            app:srcCompat="@color/green" />
+
+        <ImageView
+            android:id="@+id/learning_cat_2"
+            android:layout_width="@dimen/learning_image_size"
+            android:layout_height="@dimen/learning_image_size"
+            android:background="@color/white"
+            android:contentDescription="@string/learning_cat_2_descr"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintHorizontal_bias="0.5"
+            app:layout_constraintStart_toEndOf="@+id/learning_cat_1"
+            app:layout_constraintTop_toTopOf="parent"
+            app:srcCompat="@color/blue" />
+
+    </androidx.constraintlayout.widget.ConstraintLayout>
+</androidx.constraintlayout.widget.ConstraintLayout>
