+++ app/src/main/res/layout/fragment_main.xml
@@ -1,0 +1,41 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:context=".view.MainFragment">
+
+    <Button
+        android:id="@+id/goToDisplayDatasetButton"
+        android:layout_width="300dp"
+        android:layout_height="wrap_content"
+        android:onClick="launchToDisplayDatasetActivity"
+        android:text="@string/main_btDisplayDataset"
+        app:layout_constraintBottom_toTopOf="@+id/MainActivitySignInFragment"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@id/goToLearningButton" />
+
+    <Button
+        android:id="@+id/goToLearningButton"
+        android:layout_width="300dp"
+        android:layout_height="wrap_content"
+        android:onClick="launchToLearningActivity"
+        android:text="@string/main_btLearning"
+        app:layout_constraintBottom_toTopOf="@+id/goToDisplayDatasetButton"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
+
+    <androidx.fragment.app.FragmentContainerView
+        android:id="@+id/MainActivitySignInFragment"
+        android:name="com.github.HumanLearning2021.HumanLearningApp.view.GoogleSignInWidget"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/goToDisplayDatasetButton" />
+
+</androidx.constraintlayout.widget.ConstraintLayout>
